# custom high-performance encoder implementation