--- a/internal/proto/doc.go
+++ b/internal/proto/doc.go
@@ -1,3 +1,0 @@
-package proto
-
-//go:generate protoc --go_out=. entry.proto