This is a continuation of #49 

Now rebased to #51 (current `master`), here are the results now:
```
benchmark               old ns/op     new ns/op     delta
BenchmarkGet/128B-8     1342          935           -30.33%
BenchmarkGet/256B-8     1241          916           -26.19%
BenchmarkGet/512B-8     1502          1094          -27.16%
BenchmarkGet/1K-8       1856          1312          -29.31%
BenchmarkGet/2K-8       2554          1694          -33.67%
BenchmarkGet/4K-8       4084          2555          -37.44%
BenchmarkGet/8K-8       6694          3888          -41.92%
BenchmarkGet/16K-8      11409         6833          -40.11%
BenchmarkGet/32K-8      23304         13686         -41.27%
BenchmarkPut/128B-8     4509          4215          -6.52%
BenchmarkPut/256B-8     5815          5398          -7.17%
BenchmarkPut/1K-8       12526         11751         -6.19%
BenchmarkPut/2K-8       20990         19896         -5.21%
BenchmarkPut/4K-8       50118         36880         -26.41%
BenchmarkPut/8K-8       74180         93285         +25.75%
BenchmarkPut/16K-8      143040        149835        +4.75%
BenchmarkPut/32K-8      275379        274837        -0.20%
BenchmarkScan-8         2923          2896          -0.92%

benchmark               old MB/s     new MB/s     speedup
BenchmarkGet/128B-8     95.36        136.81       1.43x
BenchmarkGet/256B-8     206.13       279.46       1.36x
BenchmarkGet/512B-8     340.82       467.87       1.37x
BenchmarkGet/1K-8       551.65       780.14       1.41x
BenchmarkGet/2K-8       801.67       1208.85      1.51x
BenchmarkGet/4K-8       1002.76      1603.09      1.60x
BenchmarkGet/8K-8       1223.72      2106.49      1.72x
BenchmarkGet/16K-8      1435.96      2397.63      1.67x
BenchmarkGet/32K-8      1406.06      2394.13      1.70x
BenchmarkPut/128B-8     28.38        30.37        1.07x
BenchmarkPut/256B-8     44.02        47.42        1.08x
BenchmarkPut/1K-8       81.74        87.14        1.07x
BenchmarkPut/2K-8       97.57        102.93       1.05x
BenchmarkPut/4K-8       81.73        111.06       1.36x
BenchmarkPut/8K-8       110.43       87.82        0.80x
BenchmarkPut/16K-8      114.54       109.35       0.95x
BenchmarkPut/32K-8      118.99       119.23       1.00x

benchmark               old allocs     new allocs     delta
BenchmarkGet/128B-8     5              2              -60.00%
BenchmarkGet/256B-8     5              2              -60.00%
BenchmarkGet/512B-8     5              2              -60.00%
BenchmarkGet/1K-8       5              2              -60.00%
BenchmarkGet/2K-8       5              2              -60.00%
BenchmarkGet/4K-8       5              2              -60.00%
BenchmarkGet/8K-8       5              2              -60.00%
BenchmarkGet/16K-8      5              2              -60.00%
BenchmarkGet/32K-8      5              2              -60.00%
BenchmarkPut/128B-8     8              7              -12.50%
BenchmarkPut/256B-8     8              7              -12.50%
BenchmarkPut/1K-8       8              7              -12.50%
BenchmarkPut/2K-8       8              7              -12.50%
BenchmarkPut/4K-8       8              7              -12.50%
BenchmarkPut/8K-8       8              7              -12.50%
BenchmarkPut/16K-8      8              7              -12.50%
BenchmarkPut/32K-8      9              8              -11.11%
BenchmarkScan-8         35             35             +0.00%

benchmark               old bytes     new bytes     delta
BenchmarkGet/128B-8     400           168           -58.00%
BenchmarkGet/256B-8     656           296           -54.88%
BenchmarkGet/512B-8     1200          584           -51.33%
BenchmarkGet/1K-8       2288          1160          -49.30%
BenchmarkGet/2K-8       4464          2312          -48.21%
BenchmarkGet/4K-8       9072          4872          -46.30%
BenchmarkGet/8K-8       17776         9480          -46.67%
BenchmarkGet/16K-8      34928         18440         -47.21%
BenchmarkGet/32K-8      73840         40968         -44.52%
BenchmarkPut/128B-8     457           209           -54.27%
BenchmarkPut/256B-8     587           210           -64.22%
BenchmarkPut/1K-8       1459          218           -85.06%
BenchmarkPut/2K-8       2622          229           -91.27%
BenchmarkPut/4K-8       5203          249           -95.21%
BenchmarkPut/8K-8       9859          296           -97.00%
BenchmarkPut/16K-8      18897         372           -98.03%
BenchmarkPut/32K-8      41596         537           -98.71%
BenchmarkScan-8         424           424           +0.00%
```