@prologic 2019-08-07
Just for comparison; on my work Macbook Pro 13":

```
goos: darwin
goarch: amd64
pkg: github.com/prologic/bitcask
BenchmarkGet/128B-4         	 1000000	      1623 ns/op	  78.86 MB/s	     168 B/op	       2 allocs/op
BenchmarkGet/256B-4         	 1000000	      1641 ns/op	 155.98 MB/s	     296 B/op	       2 allocs/op
BenchmarkGet/512B-4         	 1000000	      1930 ns/op	 265.27 MB/s	     584 B/op	       2 allocs/op
BenchmarkGet/1K-4           	  500000	      2090 ns/op	 489.88 MB/s	    1160 B/op	       2 allocs/op
BenchmarkGet/2K-4           	  500000	      2422 ns/op	 845.53 MB/s	    2312 B/op	       2 allocs/op
BenchmarkGet/4K-4           	  500000	      2892 ns/op	1416.17 MB/s	    4872 B/op	       2 allocs/op
BenchmarkGet/8K-4           	  300000	      3823 ns/op	2142.43 MB/s	    9480 B/op	       2 allocs/op
BenchmarkGet/16K-4          	  200000	      6210 ns/op	2638.03 MB/s	   18440 B/op	       2 allocs/op
BenchmarkGet/32K-4          	  200000	     10883 ns/op	3010.88 MB/s	   40968 B/op	       2 allocs/op
BenchmarkPut/128B-4         	  200000	      8291 ns/op	  15.44 MB/s	     209 B/op	       7 allocs/op
BenchmarkPut/256B-4         	  200000	     10002 ns/op	  25.59 MB/s	     210 B/op	       7 allocs/op
BenchmarkPut/1K-4           	  100000	     21394 ns/op	  47.86 MB/s	     218 B/op	       7 allocs/op
BenchmarkPut/2K-4           	   50000	     34374 ns/op	  59.58 MB/s	     229 B/op	       7 allocs/op
BenchmarkPut/4K-4           	   20000	     66137 ns/op	  61.93 MB/s	     252 B/op	       7 allocs/op
BenchmarkPut/8K-4           	   10000	    125546 ns/op	  65.25 MB/s	     289 B/op	       7 allocs/op
BenchmarkPut/16K-4          	   10000	    212448 ns/op	  77.12 MB/s	     373 B/op	       7 allocs/op
BenchmarkPut/32K-4          	    5000	    358053 ns/op	  91.52 MB/s	     538 B/op	       8 allocs/op
BenchmarkScan-4             	 1000000	      2341 ns/op	     424 B/op	      35 allocs/op
```