@prologic 2019-08-07
That's 600k/120k read/write per second right? 🤣